requirements.txt
numpy<2.0
opencv-python-headless
Pillow
tqdm
pyyaml
matplotlib
gradio>=4.0.0
